% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodint.R
\name{prodint}
\alias{prodint}
\title{Productintegral (C++ optim)}
\usage{
prodint(A, s, t, n)
}
\arguments{
\item{A}{A function returning a matrix (intensity matrix)}

\item{s}{Initial timepoint}

\item{t}{End timepoint}

\item{n}{Number of steps for the Runge-Kutta algorithm}
}
\value{
A matrix (transition probabilities if A is an intensity matrix)
}
\description{
This function calculates the product integral of a matrix function from s to t.
It uses a Runge-Kutta method implemented in C++ for efficiency.
}
