% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_premium.R
\name{equiv_premium}
\alias{equiv_premium}
\title{Equivalence Premium
This function calculates the equivalence premium for a given insurance contract
This function requires proper construction of reward matrix like specified in the lecture notes provided in the course Liv1 at the University of Copenhagen
This function requires proper construction of the derivative of the reward matrix}
\usage{
equiv_premium(s, t, Lambda, R, dR, mu, r, n)
}
\arguments{
\item{s}{initial timepoint}

\item{t}{end timepoint}

\item{Lambda}{intensity matrix}

\item{R}{reward matrix}

\item{dR}{differential of reward matrix}

\item{mu}{equivalence premium guess}

\item{r}{constant rate as a scalar}

\item{n}{number of steps for the runge kutta algorithm}
}
\value{
a scalar
}
\description{
Equivalence Premium
This function calculates the equivalence premium for a given insurance contract
This function requires proper construction of reward matrix like specified in the lecture notes provided in the course Liv1 at the University of Copenhagen
This function requires proper construction of the derivative of the reward matrix
}
