% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodint.R
\name{prodint}
\alias{prodint}
\title{Productintegral
This function calculates the productintegral of a matrix from s to t with a given number of steps for the runge kutta}
\usage{
prodint(A, s, t, n)
}
\arguments{
\item{A}{intensity matrix}

\item{s}{initial timepoint}

\item{t}{end timepoint}

\item{n}{number of steps for the runge kutta algorithm}
}
\value{
returns a matrix (if using an intensity matrix as A you are given the transition probabilities)
}
\description{
Productintegral
This function calculates the productintegral of a matrix from s to t with a given number of steps for the runge kutta
}
\examples{
Lambda <- function(x) matrix(c(-0.1, 0.1, 0, -0.1), 2, 2)
prodint(Lambda, 0, 80, 100)
}
