% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve.R
\name{reserve}
\alias{reserve}
\title{Reserve
This function calculates the reserve given the reward matrix and some constant rate
This function requires proper construction of reward matrix like specified in the lecture notes provided in the course Liv1 at the University of Copenhagen}
\usage{
reserve(s, t, Lambda, R, mu, r, n)
}
\arguments{
\item{s}{initial timepoint}

\item{t}{end timepoint}

\item{Lambda}{intensity matrix}

\item{R}{reward matrix}

\item{mu}{equivalence premium}

\item{r}{constant rate as a scalar}

\item{n}{number of steps for the runge kutta algorithm}
}
\value{
returns an entry in a matrix
}
\description{
Reserve
This function calculates the reserve given the reward matrix and some constant rate
This function requires proper construction of reward matrix like specified in the lecture notes provided in the course Liv1 at the University of Copenhagen
}
